extract into mixin
